+++ NEWS.md
@@ -1,5 +1,6 @@
 # sigminer 1.1.0
 
+- Added `y_tr` option in `show_sig_profile()` to transform y axis values.
 - Optimized default behavior of `read_copynumber()`.
   - Support LOH records when user input minor allele copy number.
   - Set `complement = FALSE` as default.